Spline2d and its gradients are tested.

+++ python/py_diff_stokes_flow/common/common.py
@@ -41,7 +41,7 @@ class PrettyTabular(object):
 
 import shutil
 import os
-def create_folder(folder_name):
-    if os.path.isdir(folder_name):
+def create_folder(folder_name, exist_ok=False):
+    if not exist_ok and os.path.isdir(folder_name):
         shutil.rmtree(folder_name)
-    os.makedirs(folder_name)
+    os.makedirs(folder_name, exist_ok=exist_ok)
